backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Spotify Integration
spotipy==2.23.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.10.4
pydantic-settings==2.7.0

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSockets (für Live-Updates)
python-socketio==5.11.0
python-engineio==4.9.0

# Database (SQLite für Start, später Postgres)
sqlalchemy==2.0.25
alembic==1.13.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
requests==2.31.0
